--- a/src/encode.c
+++ b/src/encode.c
@@ -11,7 +11,7 @@ struct RGB {
 	float b;
 };
 
-static struct RGB multiplyBasisFunction(int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow);
+static struct RGB multiplyBasisFunction(int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow, float *cosx);
 static char *encode_int(int value, int length, char *destination);
 
 static int encodeDC(float r, float g, float b);
@@ -42,14 +42,17 @@ const char *blurHashForPixels(int xComponents, int yComponents, int width, int h
 
 	init_sRGBToLinear_cache();
 
+	float *cosx = (float *)malloc(sizeof(float) * width);
+	if (! cosx) return NULL;
 	for(int y = 0; y < yComponents; y++) {
 		for(int x = 0; x < xComponents; x++) {
-			struct RGB factor = multiplyBasisFunction(x, y, width, height, rgb, bytesPerRow);
+			struct RGB factor = multiplyBasisFunction(x, y, width, height, rgb, bytesPerRow, cosx);
 			factors[y * xComponents + x][0] = factor.r;
 			factors[y * xComponents + x][1] = factor.g;
 			factors[y * xComponents + x][2] = factor.b;
 		}
 	}
+	free(cosx);
 
 	float *dc = factors[0];
 	float *ac = dc + 3;
@@ -85,10 +88,9 @@ const char *blurHashForPixels(int xComponents, int yComponents, int width, int h
 	return destination;
 }
 
-static struct RGB multiplyBasisFunction(int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow) {
+static struct RGB multiplyBasisFunction(int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow, float *cosx) {
 	struct RGB result = { 0, 0, 0 };
 	float normalisation = (xComponent == 0 && yComponent == 0) ? 1 : 2;
-	float *cosx = (float *)malloc(sizeof(float) * width);
 
 	for(int x = 0; x < width; x++) {
 		cosx[x] = cosf(M_PI * xComponent * x / width);
@@ -104,8 +106,6 @@ static struct RGB multiplyBasisFunction(int xComponent, int yComponent, int widt
 			result.b += basis * sRGBToLinear_cache[src[3 * x + 2]];
 		}
 	}
-
-	free(cosx);
 
 	float scale = normalisation / (width * height);