Before:

```python
In [2]: %time im.size, encode(im.copy(), 9, 9)
CPU times: user 49.9 s, sys: 216 ms, total: 50.1 s
Wall time: 50.1 s
Out[2]: 
((4000, 2667),
 '|7ExFXoyM{8xM{wJTdR.NuwdMepHrE%LV[kVV@jFOUV@BBS$r@Nai_={rrR7OrRQNaMdXmw|?aRPV[t6w0x[RQkqa1xuD*xZNuWEt6n%s,Vtf+M|x[ixxswctQnPNGR-RkTJa0xtRkS~jF$%WXNHS~M|X8RjxYniR+ozs-')
```

After:

```python
In [4]: %time im.size, encode(im.copy(), 9, 9)
CPU times: user 1.42 s, sys: 0 ns, total: 1.42 s
Wall time: 1.42 s
Out[4]: 
((4000, 2667),
 '|7ExFXoyM{8xM{wJTdR.NuwdMepHrE%LV[kVV@jFOUV@BBS$r@Nai_={rrR7OrRQNaMdXmw|?aRPV[t6w0x[RQkqa1xuD*xZNuWEt6n%s,Vtf+M|x[ixxswctQnPNGR-RkTJa0xtRkS~jF$%WXNHS~M|X8RjxYniR+ozs-')
```

Also improved:

Pillow image is only closed if it was opened inside the encoder.