@homm 2024-09-24
For normal-size images speedup is also about 33x.

Before:
```
In [1]: from blurhash import encode, Image
   ...: im = Image.open('/Code/imgs/bologna.jpg').resize((120, 80), Image.BICUBIC)
   ...: assert encode(im.copy(), 6, 4) == 'W6D[Ogs.NG8xM{#m-UMepbrE%LV[I:RPBUS~soJ8nSOrM|NaMdXm'
   ...: %timeit encode(im.copy(), 6, 4)
14.6 ms ± 40.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After:
```
In [1]: from blurhash import encode, Image
   ...: im = Image.open('/Code/imgs/bologna.jpg').resize((120, 80), Image.BICUBIC)
   ...: assert encode(im.copy(), 6, 4) == 'W6D[Ogs.NG8xM{#m-UMepbrE%LV[I:RPBUS~soJ8nSOrM|NaMdXm'
   ...: %timeit encode(im.copy(), 6, 4)
433 µs ± 4.23 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```