@homm 2024-09-25
Also back ported to upstream: https://github.com/woltapp/blurhash/pull/256