@vellnes 2024-12-04
@lautat please note that
We will be very grateful for the optimization of the algorithm